refactor(ftp): restructure FtpBackend and introduce FtpCore
