+++ core/src/services/ftp/mod.rs
@@ -30,6 +30,9 @@ mod writer;
 mod backend;
 #[cfg(feature = "services-ftp")]
 pub use backend::FtpBuilder as Ftp;
+#[cfg(feature = "services-ftp")]
+mod core;
 
 mod config;
+
 pub use config::FtpConfig;
